+++ src/app.feature/auth/screen/sign-up-screen.tsx
@@ -1,9 +1,7 @@
 'use client';
 
 import {
-  AvatarImagePicker,
   Button,
-  CheckContainer,
   Dialog,
   DialogContent,
   DialogDescription,
@@ -11,24 +9,9 @@ import {
   DialogHeader,
   DialogTitle,
   Icon,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
   StepIndicator,
-  Text,
-  TextField,
-  ToggleGroup,
-  ToggleGroupItem,
 } from '@1d1s/design-system';
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormMessage,
-} from '@component/ui/form';
+import { Form } from '@component/ui/form';
 import { MEMBER_QUERY_KEYS } from '@feature/member/consts/query-keys';
 import { notifyApiError } from '@module/api/error';
 import { authStorage } from '@module/utils/auth';
@@ -38,22 +21,16 @@ import React from 'react';
 import { toast } from 'sonner';
 
 import { authApi } from '../api/auth-api';
-import {
-  SIGN_UP_GENDER_OPTIONS,
-  SIGN_UP_OCCUPATION_OPTIONS,
-  SIGN_UP_TOPIC_OPTIONS,
-  SignUpGenderValue,
-} from '../consts/sign-up-options';
 import { SignupFormValues, useSignUpForm } from '../hooks/use-sign-up-form';
 import { CategoryType, GenderType, JobType } from '../type/auth';
+import { Step1 } from './step-pages/step1';
+import { Step2 } from './step-pages/step2';
 
 type Step = 1 | 2;
 const SIGN_UP_STEPS = [
   { id: 'profile', label: '프로필 설정' },
   { id: 'topics', label: '관심 주제' },
 ];
-const SIGN_UP_LEFT_VISUAL_SIZE = 200;
-const SIGN_UP_LEFT_VISUAL_SLOT_HEIGHT = 280;
 
 function SignUpHeader({ onBack }: { onBack(): void }): React.ReactElement {
   return (
@@ -79,14 +56,6 @@ export function SignUpScreen(): React.ReactElement {
   const [step, setStep] = React.useState<Step>(1);
   const [isSubmitting, setIsSubmitting] = React.useState(false);
   const [showExitDialog, setShowExitDialog] = React.useState(false);
-  const [imgPreviewUrl, setImgPreviewUrl] = React.useState<
-    string | undefined
-  >();
-  const currentYear = new Date().getFullYear();
-  const yearOptions = Array.from({ length: 100 }, (_, i) => currentYear - i);
-  const monthOptions = Array.from({ length: 12 }, (_, i) => i + 1);
-  const dayOptions = Array.from({ length: 31 }, (_, i) => i + 1);
-  const selectedTopics = form.watch('topics') ?? [];
 
   const onSubmit = async (values: SignupFormValues): Promise<void> => {
     const accessToken = authStorage.getAccessToken();
@@ -219,419 +188,13 @@ export function SignUpScreen(): React.ReactElement {
           </div>
 
           {step === 1 ? (
-            <div className="mx-auto flex w-full max-w-[1080px] flex-1 items-stretch px-4 pb-5">
-              <div className="grid w-full gap-6 lg:min-h-0 lg:grid-cols-[0.9fr_1.1fr]">
-                <section className="flex min-h-0 flex-col pt-2 text-left">
-                  <Text size="display2" weight="bold" className="text-gray-900">
-                    나에 대해 알려주세요.
-                  </Text>
-                  <Text
-                    size="body2"
-                    weight="regular"
-                    className="mt-3 text-gray-600"
-                  >
-                    몇 가지 정보를 입력하고 맞춤 챌린지를 추천받아보세요.
-                  </Text>
-
-                  <div className="mt-6 flex justify-center lg:mt-8">
-                    <FormField
-                      control={form.control}
-                      name="img"
-                      render={({ field }) => (
-                        <FormItem className="mb-5">
-                          <AvatarImagePicker
-                            size={SIGN_UP_LEFT_VISUAL_SIZE}
-                            defaultImageUrl={imgPreviewUrl}
-                            changeLabel="사진 추가"
-                            onChange={(
-                              event: React.ChangeEvent<HTMLInputElement>
-                            ) => {
-                              const file = event.target.files?.[0] || undefined;
-                              field.onChange(file);
-                              if (file) {
-                                setImgPreviewUrl(URL.createObjectURL(file));
-                              }
-                            }}
-                          />
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-                  </div>
-                </section>
-
-                <section className="rounded-4 flex min-h-0 flex-col border border-gray-200 bg-white p-5 shadow-[0_8px_20px_rgba(34,34,34,0.04)] lg:max-h-[620px] lg:self-start lg:overflow-y-auto lg:p-6">
-                  <FormField
-                    control={form.control}
-                    name="nickname"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormControl>
-                          <TextField
-                            label="닉네임"
-                            placeholder="예: 챌린저123"
-                            id="nickname"
-                            className="w-full"
-                            value={field.value ?? ''}
-                            onChange={field.onChange}
-                            onBlur={field.onBlur}
-                            name={field.name}
-                            ref={field.ref}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <div className="mt-5">
-                    <Text
-                      size="body2"
-                      weight="bold"
-                      className="mb-1 text-gray-800"
-                    >
-                      생년월일
-                    </Text>
-                    <div className="grid grid-cols-3 gap-3">
-                      <FormField
-                        control={form.control}
-                        name="year"
-                        render={({ field }) => (
-                          <FormItem>
-                            <Select
-                              value={field.value}
-                              onValueChange={field.onChange}
-                            >
-                              <FormControl>
-                                <SelectTrigger className="w-full !min-w-0">
-                                  <SelectValue placeholder="년" />
-                                </SelectTrigger>
-                              </FormControl>
-                              <SelectContent>
-                                {yearOptions.map((year) => (
-                                  <SelectItem key={year} value={String(year)}>
-                                    {year}년
-                                  </SelectItem>
-                                ))}
-                              </SelectContent>
-                            </Select>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="month"
-                        render={({ field }) => (
-                          <FormItem>
-                            <Select
-                              value={field.value}
-                              onValueChange={field.onChange}
-                            >
-                              <FormControl>
-                                <SelectTrigger className="w-full !min-w-0">
-                                  <SelectValue placeholder="월" />
-                                </SelectTrigger>
-                              </FormControl>
-                              <SelectContent>
-                                {monthOptions.map((month) => (
-                                  <SelectItem key={month} value={String(month)}>
-                                    {month}월
-                                  </SelectItem>
-                                ))}
-                              </SelectContent>
-                            </Select>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-
-                      <FormField
-                        control={form.control}
-                        name="day"
-                        render={({ field }) => (
-                          <FormItem>
-                            <Select
-                              value={field.value}
-                              onValueChange={field.onChange}
-                            >
-                              <FormControl>
-                                <SelectTrigger className="w-full !min-w-0">
-                                  <SelectValue placeholder="일" />
-                                </SelectTrigger>
-                              </FormControl>
-                              <SelectContent>
-                                {dayOptions.map((day) => (
-                                  <SelectItem key={day} value={String(day)}>
-                                    {day}일
-                                  </SelectItem>
-                                ))}
-                              </SelectContent>
-                            </Select>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-                  </div>
-
-                  <div className="mt-5">
-                    <Text
-                      size="body2"
-                      weight="bold"
-                      className="mb-1 text-gray-800"
-                    >
-                      성별
-                    </Text>
-                    <FormField
-                      control={form.control}
-                      name="gender"
-                      render={({ field }) => (
-                        <FormItem>
-                          <ToggleGroup
-                            type="single"
-                            value={field.value}
-                            onValueChange={(value) => {
-                              if (value) {
-                                field.onChange(value as SignUpGenderValue);
-                              }
-                            }}
-                            className="grid grid-cols-3 gap-2"
-                          >
-                            {SIGN_UP_GENDER_OPTIONS.map((option) => (
-                              <ToggleGroupItem
-                                key={option.value}
-                                value={option.value}
-                                shape="square"
-                                className="h-10 w-full justify-center px-0"
-                              >
-                                {option.label}
-                              </ToggleGroupItem>
-                            ))}
-                          </ToggleGroup>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-                  </div>
-
-                  <div className="mt-5">
-                    <Text
-                      size="body2"
-                      weight="bold"
-                      className="mb-1 text-gray-800"
-                    >
-                      직업
-                    </Text>
-                    <FormField
-                      control={form.control}
-                      name="job"
-                      render={({ field }) => (
-                        <FormItem>
-                          <Select
-                            value={field.value}
-                            onValueChange={field.onChange}
-                          >
-                            <FormControl>
-                              <SelectTrigger className="w-full !min-w-0">
-                                <SelectValue placeholder="직업 선택" />
-                              </SelectTrigger>
-                            </FormControl>
-                            <SelectContent>
-                              {SIGN_UP_OCCUPATION_OPTIONS.map((option) => (
-                                <SelectItem
-                                  key={option.value}
-                                  value={option.value}
-                                >
-                                  {option.label}
-                                </SelectItem>
-                              ))}
-                            </SelectContent>
-                          </Select>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-                  </div>
-
-                  <div className="mt-5">
-                    <Text
-                      size="body2"
-                      weight="bold"
-                      className="mb-1 text-gray-800"
-                    >
-                      프로필 공개
-                    </Text>
-                    <FormField
-                      control={form.control}
-                      name="isPublic"
-                      render={({ field }) => (
-                        <FormItem>
-                          <ToggleGroup
-                            type="single"
-                            value={field.value ? 'public' : 'private'}
-                            onValueChange={(value) => {
-                              if (value) {
-                                field.onChange(value === 'public');
-                              }
-                            }}
-                            className="grid grid-cols-2 gap-2"
-                          >
-                            <ToggleGroupItem
-                              value="public"
-                              shape="square"
-                              className="h-10 w-full justify-center px-0"
-                            >
-                              공개
-                            </ToggleGroupItem>
-                            <ToggleGroupItem
-                              value="private"
-                              shape="square"
-                              className="h-10 w-full justify-center px-0"
-                            >
-                              비공개
-                            </ToggleGroupItem>
-                          </ToggleGroup>
-                        </FormItem>
-                      )}
-                    />
-                  </div>
-
-                  <div className="mt-8 flex justify-end">
-                    <Button
-                      type="button"
-                      size="medium"
-                      onClick={handleNextStep}
-                    >
-                      다음 단계
-                    </Button>
-                  </div>
-                </section>
-              </div>
-            </div>
+            <Step1 onNext={handleNextStep} />
           ) : (
-            <div className="mx-auto flex w-full max-w-[1080px] flex-1 items-stretch px-4 pb-5">
-              <div className="grid w-full gap-6 lg:min-h-0 lg:grid-cols-[0.9fr_1.1fr]">
-                <section className="flex min-h-0 flex-col pt-2 text-left">
-                  <Text size="display2" weight="bold" className="text-gray-900">
-                    어떤 주제에 관심이 있나요?
-                  </Text>
-                  <Text
-                    size="body2"
-                    weight="regular"
-                    className="mt-3 text-gray-600"
-                  >
-                    도전하고 싶은 관심 주제를 선택해주세요.
-                  </Text>
-                  <div
-                    aria-hidden
-                    className="mt-6 hidden lg:mt-8 lg:block"
-                    style={{ height: SIGN_UP_LEFT_VISUAL_SLOT_HEIGHT }}
-                  />
-                </section>
-
-                <section className="rounded-4 flex min-h-0 flex-col border border-gray-200 bg-white p-5 shadow-[0_8px_20px_rgba(34,34,34,0.04)] lg:max-h-[620px] lg:self-start lg:overflow-y-auto lg:p-6">
-                  <FormField
-                    control={form.control}
-                    name="topics"
-                    render={({ field }) => (
-                      <FormItem>
-                        <Text
-                          size="body2"
-                          weight="regular"
-                          className="mb-3 text-gray-500"
-                        >
-                          최대 3개까지 선택할 수 있어요.
-                        </Text>
-                        <div className="grid grid-cols-2 gap-3 lg:grid-cols-3">
-                          {SIGN_UP_TOPIC_OPTIONS.map((option) => {
-                            const checked = Array.isArray(selectedTopics)
-                              ? selectedTopics.includes(option.value)
-                              : false;
-                            const isMaxReached = selectedTopics.length >= 3;
-
-                            return (
-                              <CheckContainer
-                                key={option.value}
-                                type="button"
-                                checked={checked}
-                                disabled={!checked && isMaxReached}
-                                onCheckedChange={(nextChecked) => {
-                                  const currentTopics = Array.isArray(
-                                    form.getValues('topics')
-                                  )
-                                    ? form.getValues('topics')
-                                    : [];
-                                  if (
-                                    nextChecked &&
-                                    currentTopics.length >= 3
-                                  ) {
-                                    return;
-                                  }
-                                  const nextTopics = nextChecked
-                                    ? Array.from(
-                                        new Set([
-                                          ...currentTopics,
-                                          option.value,
-                                        ])
-                                      )
-                                    : currentTopics.filter(
-                                        (topic) => topic !== option.value
-                                      );
-
-                                  field.onChange(nextTopics);
-                                }}
-                                width="100%"
-                                height={96}
-                                showCheckIndicator
-                                className="rounded-3 w-full min-w-0 px-3"
-                              >
-                                <div className="flex w-full flex-col items-center justify-center gap-2">
-                                  <span className="text-xl leading-none">
-                                    {option.icon}
-                                  </span>
-                                  <Text
-                                    size="body2"
-                                    weight="medium"
-                                    className="text-gray-700"
-                                  >
-                                    {option.label}
-                                  </Text>
-                                </div>
-                              </CheckContainer>
-                            );
-                          })}
-                        </div>
-                        <div className="mt-4 text-left">
-                          <FormMessage />
-                        </div>
-                      </FormItem>
-                    )}
-                  />
-
-                  <div className="mt-8 flex items-center justify-between gap-3">
-                    <Button
-                      type="button"
-                      size="medium"
-                      variant="outlined"
-                      disabled={isSubmitting}
-                      onClick={() => setStep(1)}
-                    >
-                      이전 단계
-                    </Button>
-                    <Button
-                      type="button"
-                      size="medium"
-                      disabled={isSubmitting}
-                      onClick={() => void form.handleSubmit(onSubmit)()}
-                    >
-                      {isSubmitting ? '처리 중...' : '가입 완료'}
-                    </Button>
-                  </div>
-                </section>
-              </div>
-            </div>
+            <Step2
+              onPrev={() => setStep(1)}
+              onSubmit={() => void form.handleSubmit(onSubmit)()}
+              isSubmitting={isSubmitting}
+            />
           )}
         </form>
       </Form>
